move message subscriber to ToolkitConnectionManager
